Added CNN preprocess
Added Jupyter Notebook where I preprocessed 'CNN_Articles_sentiment.tsv' by @AhmedSabanovic. In the preprocessing the most relevant words of each article are extracted, and then aggregated by year. In the aggregation step the frequency of each word per year is calculated, and the average sentiment of that word in the respective year is calculated.