Added preprocessed CNN dataset
Added preprocessed CNN dataset in the desired format, as requested by @AhmedSabanovic. The most relevant words of each article are extracted, and then aggregated by year. In the aggregation step the frequency of each word per year is calculated, and the average sentiment of that word in the respective year is calculated.